requirements.txt
cryptography==3.4.8
PyJWT==2.3.0
requests==2.25.1